generalize file download fix to apply to transcripts
